refactor(graduate): config 파일 구조 개선 및 상수 정리
