+++ apps/graduate/src/admin/widgets/Table/model/useAdminDownload.ts
@@ -2,14 +2,15 @@ import { useCallback } from 'react';
 
 import { GraduationType } from '~/shared/constants';
 import { useToast } from '~/shared/hooks';
-import { downloadGraduationUsersExcel } from '~/shared/utils';
+
+import { excelDownload } from '../utils/excelDownload';
 
 export function useAdminDownload(graduationType?: GraduationType) {
   const { toast } = useToast();
 
   const handleDownload = useCallback(async () => {
     try {
-      await downloadGraduationUsersExcel(graduationType);
+      await excelDownload(graduationType);
     } catch (error) {
       toast.error(
         error instanceof Error ? error.message : '다운로드에 실패했습니다.',
