+++ apps/graduate/src/client/pages/home/ui/NoticeSection.tsx
@@ -2,7 +2,8 @@ import { Link, useNavigate } from '@tanstack/react-router';
 import { ChevronRight, FileText } from 'lucide-react';
 
 import { EmptyState } from '~/shared/components';
-import { ROUTE, PAGINATION, EMPTY_MESSAGES } from '~/shared/constants';
+import { PAGINATION, EMPTY_MESSAGES } from '~/shared/config';
+import { ROUTE } from '~/shared/constants';
 import { useNoticeList } from '~/shared/hooks';
 
 import * as styles from '../styles/HomePage.css';
