+++ apps/graduate/src/shared/config/carousel.ts
@@ -1,0 +1,9 @@
+/**
+ * 캐러셀 설정
+ */
+export const CAROUSEL_CONFIG = {
+  /** 자동 재생 속도 (밀리초) */
+  AUTOPLAY_SPEED: 5000,
+  /** 페이드 효과 지속 시간 (밀리초) */
+  FADE_DURATION: 300,
+} as const;
