+++ apps/graduate/src/shared/config/colors.ts
