+++ apps/graduate/src/shared/config/index.ts
@@ -1,1 +1,6 @@
+export * from './carousel';
+export * from './colors';
+export * from './date';
 export * from './modal';
+export * from './pagination';
+export * from './statusMessage';
