+++ apps/graduate/src/shared/config/pagination.ts
@@ -1,0 +1,13 @@
+/**
+ * 페이지네이션 기본 설정
+ */
+export const PAGINATION = {
+  /** 고정 공지사항 표시 개수 */
+  PINNED_NOTICE_SIZE: 3,
+  /** 공지사항 목록 기본 페이지 크기 */
+  NOTICE_LIST_SIZE: 10,
+  /** 일정 목록 기본 페이지 크기 */
+  SCHEDULE_LIST_SIZE: 5,
+  /** 최근 공지사항 표시 개수 */
+  RECENT_NOTICE_SIZE: 5,
+} as const;
