+++ apps/graduate/src/shared/constants/graduationWorkflow.ts
@@ -3,6 +3,7 @@
  */
 
 import { WORKFLOW_STAGE, type WorkflowStage } from '~/shared/types/graduation';
+
 import ROUTE from './route';
 
 /**
