+++ apps/graduate/src/shared/constants/index.ts
@@ -1,12 +1,17 @@
 export * from './professors';
 export * from './endpoint';
 export * from './graduationWorkflow';
-export * from './date';
+export * from '../config/date';
 export * from './keys';
 export * from './schedule';
-export * from './colors';
-export * from './ui';
 export { default as ROUTE } from './route';
 
 // Re-export graduation types and constants from types
-export { GRADUATION_TYPE, USER_ROLE, APPROVAL_STATUS, type GraduationType, type UserRole, type ApprovalStatus } from '../types/graduation';
+export {
+  GRADUATION_TYPE,
+  USER_ROLE,
+  APPROVAL_STATUS,
+  type GraduationType,
+  type UserRole,
+  type ApprovalStatus,
+} from '../types/graduation';
