+++ apps/graduate/src/shared/utils/index.ts
@@ -5,4 +5,3 @@ export { default as parseError } from './error';
 export { default as logout } from './logout';
 export { default as queryClient } from './queryClient';
 export { default as router } from './router';
-export { default as downloadGraduationUsersExcel } from './downloadGraduationUsersExcel';
